agents/coverage/requirements.txt
# Microsoft Agent Framework — core runtime
agent-framework-core>=1.0.0rc2,<=1.0.0rc3
# Foundry Hosted Agent server adapter — core protocol layer + MAF adapter
# See: https://learn.microsoft.com/azure/foundry/agents/concepts/hosted-agents#public-adapter-packages
azure-ai-agentserver-core>=1.0.0b16
azure-ai-agentserver-agentframework>=1.0.0b16
# azure-ai-projects is pulled transitively by agent-framework-core
# Managed identity / DefaultAzureCredential
azure-identity>=1.15.0
# Environment variable loading
python-dotenv>=1.0.0
# Azure Monitor OpenTelemetry (optional — remove if not using App Insights)
azure-monitor-opentelemetry>=1.6.0
